Major perf boost for day 5
A stack lets us run the whole reaction in O(N) time, which on my machine, with Python, means 8 milliseconds. That's a lot better than the 2 seconds for a regex approach. Part 2, building on top of the work from part 1 is down to 54-ish milliseconds, down from nearly a full minute.